Add Product Screen and add title for each tab
Creates website for products. Each product after clicked will show its
own picture of the product on the left, and all the info on the right
with its current status (available or unavailable). If it is available,
then the add to cart button will show. If it is not available (out of
order), then there will not be add to cart button.
Also added title for each tab name. If on homepage, the tab will show
eBruin. On product page the tab will show product name.