Add Cart Functionality
Create cart on the header for all screens, and whenever the add to cart
button is clicked, the quantity will appear next to cart on top. The
number indicates the total quantity of the items, it can be only one
item with multiply amount. Not the same as outputting the number of
items in the cart. e.g. can show 10 but only 2 items.