Shopping Cart Screen
Created CartScreen.js with all the fronend implementation inside. Also
add handlers to all the buttons associated with this screen. Also inside
Store.js set all the cart action type to local storage so that if the
user refresh the page the items in the cart will still exist. Added
event handler in home screen for add to cart button, and create check
out button handler to navigate to sign in page once clicked.